created leaflet map with markers and circle
Loading markers from JSON data in data/marker_coordinates.json
Marker coordinates all in nearby neighborhoods
Using one coordinate and radius (miles) as parameters for a circle
This circle represents a driver's desired rider pick up area.
If you change the value of the radius via the input form,
the size of the circle changes with it.