added driver map ui
Created a leaflet map with some sample markers loaded onto it.
It creates a circle with the center where the user specified they live.
By default the circle has a radius of 5 miles (easy to change).

When the user drags the circle around the map, or changes the radius
of the circle, this information is automatically persisted in the database